:tada: Update button variant and colors
