+++ @navikt/core/react/src/button/Button.tsx
@@ -7,6 +7,14 @@ import { omit } from "../util";
 import { composeEventHandlers } from "../util/composeEventHandlers";
 import { OverridableComponent } from "../util/types";
 
+type legacyVariants =
+  | "primary-neutral"
+  | "secondary-neutral"
+  | "tertiary-neutral"
+  | "danger";
+
+type HiddenVariant = legacyVariants & { __brand?: never };
+
 export interface ButtonProps
   extends React.ButtonHTMLAttributes<HTMLButtonElement> {
   /**
@@ -17,14 +25,7 @@ export interface ButtonProps
    * Changes design and interaction-visuals.
    * @default "primary"
    */
-  variant?:
-    | "primary"
-    | "primary-neutral"
-    | "secondary"
-    | "secondary-neutral"
-    | "tertiary"
-    | "tertiary-neutral"
-    | "danger";
+  variant?: "primary" | "secondary" | "tertiary" | HiddenVariant;
   /**
    * Changes padding, height, and font-size.
    * @default "medium"
@@ -50,6 +51,10 @@ export interface ButtonProps
    * @default "left"
    */
   iconPosition?: "left" | "right";
+  /**
+   * Button color
+   */
+  "data-color"?: AkselColor;
 }
 
 /**
@@ -100,18 +105,12 @@ export const Button: OverridableComponent<ButtonProps, HTMLButtonElement> =
           {...filterProps}
           ref={ref}
           onKeyUp={composeEventHandlers(onKeyUp, handleKeyUp)}
-          className={cn(
-            className,
-            "navds-button",
-            `navds-button--${variant}`,
-            `navds-button--${size}`,
-            {
-              "navds-button--loading": loading,
-              "navds-button--icon-only": !!icon && !children,
-              "navds-button--disabled": disabled ?? loading,
-            },
-          )}
-          disabled={disabled ?? loading ? true : undefined}
+          className={cn(className, "navds-button", `navds-button--${size}`, {
+            "navds-button--loading": loading,
+            "navds-button--icon-only": !!icon && !children,
+            "navds-button--disabled": disabled ?? loading,
+          })}
+          disabled={(disabled ?? loading) ? true : undefined}
         >
           {icon && iconPosition === "left" && (
             <span className={cn("navds-button__icon")}>{icon}</span>
